WIP: Move base Tx types to public header
