+++ src/js/wrap.h
@@ -3,9 +3,9 @@
 #pragma once
 
 #include "ccf/historical_queries_interface.h"
+#include "ccf/tx.h"
 #include "ds/logger.h"
 #include "kv/kv_types.h"
-#include "kv/tx.h"
 
 #include <memory>
 #include <quickjs/quickjs-exports.h>
