+++ src/node/ledger_secret.h
@@ -5,7 +5,6 @@
 #include "crypto/entropy.h"
 #include "crypto/symmetric_key.h"
 #include "kv/kv_types.h"
-#include "kv/tx.h"
 #include "secrets.h"
 #include "shares.h"
