+++ src/node/ledger_secrets.h
@@ -2,10 +2,10 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/tx.h"
 #include "crypto/entropy.h"
 #include "crypto/symmetric_key.h"
 #include "kv/kv_types.h"
-#include "kv/tx.h"
 #include "ledger_secret.h"
 #include "secrets.h"
 #include "shares.h"
